update system time after N900 resumes
Currently nothing updates the system clock after a suspend to memory
operation, and date shows that almost no time elapsed.
As ARM doesn't have a specific read_persistent_clock the default
returns 0.  The RTC rtc_suspend/rtc_resume will set the clock after
resume, but requires CONFIG_RTC_HCTOSYS_DEVICE to be set, but that
requires RTC to be compiled into the kernel.

This splits up the configuration for setting the system time on bootup
from setting the time on resume.  This allows RTC_CLASS to be compiled
as a module if the system needs the RTC for resume, but not for boot,
which is the case for the N900.