musb resume fix to charge/talk to computer USB port
Without this patch series once the N900 was suspend to ram and resumed
it would not be able to charge or be detected when plugged into a
computer USB port.  This fixes the registers by using the previously
stored registers rather than the registers at the time of suspend.

There are other fixes such as actually sleeping when probing for the
charger rather than busy waiting.