Add patch nokia-av_key.patch
With this patch pressed headset button will be reported immediately as KEY_PROG1.
But sometimes this key can be reported also when was not presses.

Application should take care about false positive pressing, so this is reason
why kernel will report it as KEY_PROG1 (and not KEY_PHONE).

To verify if key was pressed application can wait for real KEY_PHONE event which
would be delivered after few seconds as before.

Application can receive notify about this presses key via HAL "prog1" key.
So this will not break any existing applications.

Thanks Pavel Machek for it!